\name{(INTRO.DISTRTEST)}
\alias{(INTRO.DISTRTEST)}
\title{DISTRTEST}
\center{Estimation and Testing classes based on package distr}
\synopsis{}

\section{Usage}{
   library("distrTEst") 
}


\description{
\pkg{``distrTest''} is a package for R from version 2.2.0 onwards that is distributed
under GPL license 2.0. 
It requires package \pkg{setRNG} by Paul Gilbert, \email{pgilbert@bank-banque-canada.ca},
and packages \pkg{distr} and \pkg{distrSim} by the authors of this package, 
\email{peter.ruckdeschel@uni-bayreuth.de}, to be installed from 

\url{http://cran.r-project.org/mirrors.html}

The aim of this package is to provide a conceptual framework
to treat the evaluation of statistical procedures (estimators/tests) 
at data/simulations in a unified way. This is achieved by means of the S4 class
Evaluation. 
The package is based on our packages \pkg{distr} and \pkg{distrSim}, hence uses distribution classes 
and simulation classes as introduced there to capture the situation from which the simulations stem.

A somewhat longer and more detailed manual can be obtained under

\url{http://www.uni-bayreuth.de/departments/math/org/mathe7/DISTR/}

}
\keyword{}
