\name{call.ev-methods}
\docType{methods}
\alias{call.ev-methods}
\alias{call.ev}
\alias{call.ev,Evaluation-method}
\title{ Methods for Function call.ev in Package `distrTEst' }
\description{call.ev-methods}
\section{Methods}{\describe{
    \item{call.ev}{\code{signature(object = "Evaluation")}: returns the call which created the object }
}}
\keyword{methods}
\concept{S4 evaluation class}
\concept{accessor function}
\concept{call of evaluation}
