\name{Data-methods}
\docType{methods}
\alias{Data-methods}
\alias{Data}
\alias{Data,Evaluation-method}
\alias{Data,EvaluationList-method}

\title{ Methods for Function Data in Package `distrTEst'}
\description{Data-methods}
\section{Methods}{\describe{
\item{Data}{\code{signature(object = "Evaluation")}: returns the Data slot }
\item{Data}{\code{signature(object = "EvaluationList")}: returns the common Data slot of
      the respective list elements}
}}
\keyword{methods}
\concept{S4 evaluation class}
\concept{accessor function}
