\name{internal-functions}
\alias{distrTEst-internals}
\alias{.convert.result.format}
\title{Internal routines/objects in Package `distrTEst' }
\description{internal routines/objects in Package `distrTEst' }

\usage{
.convert.result.format(x, resname="res")
}
\arguments{
  \item{x}{the result object to be converted}
  \item{resname}{the default-name for the column(s) of the result}
}
\details{\code{.convert.result.format} returns the result
converted to a data.frame of dimensions runs x resultDim}

\author{
  Peter Ruckdeschel \email{Peter.Ruckdeschel@uni-bayreuth.de}
  }

\keyword{internal}
\keyword{misc}
\concept{conversion}