\name{estimator-methods}
\docType{methods}
\alias{estimator-methods}
\alias{estimator}
\alias{estimator,Evaluation-method}

\title{ Methods for Function estimator in Package `distrTEst'}
\description{estimator-methods}
\section{Methods}{\describe{

\item{estimator}{\code{signature(object = "Evaluation")}: returns the estimator }
}}
\concept{accessor function}
\concept{S4 evaluation class}
\keyword{methods}
