\name{evaluate-methods}
\docType{methods}
\alias{evaluate-methods}
\alias{evaluate}
\alias{evaluate,Dataclass,function-method}
\title{ Methods for Function evaluate in Package `distrTEst'}
\description{evaluate-methods}
\section{Methods}{\describe{
    \item{evaluate}{\code{signature(object = "Dataclass", estimator = "function")}:

      creates an object of type "Evaluation", see there for further
      information}}}
\seealso{\code{\link{Evaluation-class}}}
\keyword{}
