\name{filename-methods}
\docType{methods}
\alias{filename-methods}
\alias{filename}
\alias{filename,Evaluation-method}

\title{ Methods for Function filename in Package `distrTEst' }
\description{filename-methods}
\section{Methods}{\describe{
   
\item{filename}{\code{signature(object = "Evaluation")}: returns the filename of the evaluated object }
}}
\keyword{file}
\concept{S4 evaluation class}
\keyword{methods}
\concept{accessor function}
