\name{name-methods}
\docType{methods}
\alias{name-methods}
\alias{name}
\alias{name,Evaluation-method}
\title{ Methods for Function name in Package `distrTEst' }
\description{name-methods}
\section{Methods}{\describe{
\item{name}{\code{signature(object = "Evaluation")}: returns the slot name of data object }
}}
\keyword{}
