\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,Evaluation-method}
\title{ Methods for Function plot in Package `distrTEst' }
\description{plot-methods}
\section{Methods}{\describe{
\item{plot}{\code{signature(object = "Evaluation")}: returns a boxplot of the result }
}}
\keyword{}
