\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,Evaluation-method}
\title{ Methods for Function print in Package `distrTEst' }
\description{print-methods}

\section{Methods}{\describe{

\item{print}{\code{signature(x = "Evaluation")}: returns the name of the data object, its filename, the estimator used and the result }
}}
\keyword{}
