\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,Evaluation-method}
\title{ Methods for Function summary in Package `distrTEst' }
\description{summary-methods}

\section{Methods}{\describe{


\item{summary}{\code{signature(object = "Evaluation")}: returns the name of the data object, its filename, the estimator used and
a statistical summary of the result }

}}
\keyword{}
