\name{name-methods}
\docType{methods}
\alias{name-methods}
\alias{name}
\alias{name,Evaluation-method}
\title{ Methods for Function name in Package `distrTEst' }
\description{name-methods}
\section{Methods}{\describe{
\item{name}{\code{signature(object = "Evaluation")}: returns the slot name of data object }
\item{name<-}{\code{signature(.Object = "Evaluation")}: modifies the slot name of data object }
}}
\concept{S4 evaluation class}
\keyword{methods}
\concept{accessor function}
