\name{numericorNULL-class}
\docType{class}
\alias{numericorNULL-class}

\title{Class "numericorNULL" }

\description{ auxiliary class; may contain either a numeric vector or NULL.}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "numericorNULL" in the signature.
}

\author{
  Thomas Stabla \email{statho3@web.de},\cr 
  Florian Camphausen \email{fcampi@gmx.de},\cr
  Peter Ruckdeschel \email{Peter.Ruckdeschel@uni-bayreuth.de},\cr 
  Matthias Kohl \email{Matthias.Kohl@stamats.de}
  }

\note{ Dataclass-class can save data either of type "NULL" (means no data) or "vector"}

\seealso{\code{\link[distrSim]{Dataclass-class}}}

\keyword{}
