\name{numericorNULL-class}
\docType{class}
\alias{numericorNULL-class}
\alias{DataframeorNULL-class}
\alias{CallorNULL-class}

\title{Classes "numericorNULL", "CallorNULL", and "DataframeorNULL"}

\description{ auxiliary classes; may contain either a numeric vector or NULL 
[or a call / data.frame or NULL, respectively].}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "numericorNULL", "CallorNULL", and "DataframeorNULL" in the signature.
}

\author{
  Thomas Stabla \email{statho3@web.de},\cr 
  Florian Camphausen \email{fcampi@gmx.de},\cr
  Peter Ruckdeschel \email{Peter.Ruckdeschel@web.de},\cr 
  Matthias Kohl \email{Matthias.Kohl@stamats.de}
  }

\note{From version 1.8, the result slot of an object of class evaluation is of type "DataframeorNULL"}

\seealso{\code{\link{Evaluation-class}}}

\keyword{classes}
\concept{class union}
\concept{virtual}
